Backend/requirements.txt
# Google Gen AI SDK & ADK (Vertex AI)
google-genai>=1.0.0
google-adk>=0.3.0
google-auth>=2.27.0
google-cloud-aiplatform>=1.40.0

# PostgreSQL & pgvector
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# WebSocket server
websockets>=12.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Async utilities
aiohttp>=3.9.0
httpx>=0.26.0

# Environment & configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Date/time handling
pytz>=2024.1

# Logging
structlog>=24.1.0

# Testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
